refactored the names to adhere to the agreed upon standards
