+++ client/src/Components/CountryBorder/CountryBorder.less
@@ -1,4 +1,4 @@
-.Country_Border {
+.CountryBorder {
   display: flex;
   flex-direction: column;
   justify-content: center;
@@ -6,12 +6,12 @@
   border: 1px solid black;
   // overflow: hidden;
 
-  .Country_Border-Border {
+  .CountryBorder__Border {
     text-align: center;
     width: 50%;
     border: 1px solid black;
   }
-  .Country_Border-Slider {
+  .CountryBorder__Slider {
     position: relative;
     z-index: 10;
